refactor: Entity 확장 함수 이동
