+++ app/src/main/java/com/into/websoso/domain/mapper/NoticeMapper.kt
@@ -1,11 +1,9 @@
 package com.into.websoso.domain.mapper
 
+import com.into.websoso.data.model.NotificationEntity
 import com.into.websoso.data.model.NotificationsEntity
-import com.into.websoso.data.model.NotificationsEntity.NotificationEntity
 import com.into.websoso.domain.model.Notice
 import com.into.websoso.domain.model.NoticeInfo
-import com.into.websoso.domain.model.NoticeType
-import com.into.websoso.domain.usecase.GetNoticeListUseCase.Companion.DEFAULT_INTRINSIC_ID
 
 fun NotificationsEntity.toDomain(): NoticeInfo =
     NoticeInfo(
@@ -25,18 +23,3 @@ fun NotificationEntity.toDomain(): Notice =
         isRead = isRead,
         intrinsicId = getIntrinsicId(),
     )
-
-private fun NotificationEntity.getNoticeType(): NoticeType =
-    NoticeType.from(
-        when {
-            isNotice -> "NOTICE"
-            else -> "FEED"
-        },
-    )
-
-private fun NotificationEntity.getIntrinsicId(): Long =
-    when {
-        isNotice -> notificationId
-        feedId != null -> feedId
-        else -> DEFAULT_INTRINSIC_ID
-    }
